Small fixes to URL handling in links.
By adding a couple of options, links are no longer converted with
relative URLs when referencing contents of the same site (we need this
a lot where I work, for example in archive pages).
Found a small typo that caused a JS error when a page had a
non-image attachment.
I don't know why indentation was so broken, but I fixed that too.